+++ dahora_app/handlers/quit_app_handler.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Optional
 from dahora_app.callback_manager import CallbackHandler
 
 if TYPE_CHECKING:
-    from dahora_app import DahoraApp  # Import apenas para type hints
+    from main import DahoraApp  # Import apenas para type hints
 
 
 logger = logging.getLogger(__name__)